+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   available regardless of whether the `global_allocator` feature is enabled. The
   `global_allocator` feature now only controls whether `allocator::Allocator` is
   set as Rust's global allocator.
+- `Error::new` and `Error::from` now panic if the status is `SUCCESS`.
 
 ## uefi-macros - [Unreleased]